Add style and main menu actions
I created and filled out the css in particular I'm going for a sleek
design. Maybe not the best for a recipe site but a good exercise.
In particular I added an effect to the links on the main menu that
dims the other links on hover and pops up an image. I don't know for
sure how it works, but it does.